+++ core/src/main/java/com/github/steanky/proxima/path/PathOperation.java
@@ -4,7 +4,7 @@ import org.jetbrains.annotations.NotNull;
 
 public interface PathOperation {
     void init(int startX, int startY, int startZ, int destinationX, int destinationY, int destinationZ,
-            @NotNull PathSettings settings, double yOffset);
+            @NotNull PathSettings settings, float yOffset);
 
     boolean step();
